\name{fastplm-internal}
\alias{create.indicators}
\alias{create.subindicators}
\alias{create.complex.effect}
\alias{create.fixed.effects}
\alias{fastplm.core}
\alias{SolveFixedEffects}


\title{Internal Fastplm Functions}
\description{
 Internal fastplm functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ts}
