\name{fastplm-package}
\alias{fastplm-package}
\docType{package}
\title{
    A package for fast fixed-effects algorithms.
}
\description{
    Fast Estimation of Linear Model and 2SLS Model with Multiple Fixed Effects
}
\details{This package performs fast estimation of multiple levels of fixed effects and various standard errors(clustered, jackknife and bootstrap).}

\author{
    Yiqing Xu(Maintainer); Minsheng Liu; Licheng Liu; Ziyi Liu
}

\references{
Anderson, T. W. and H. Rubin. 1949. Estimation of the parameters of a single equation in a complete system of stochastic equations. Annals of Mathematical Statistics, Vol. 20, pp. 46-63.

Angrist, J.D. and Pischke, J.-S. 2009. Mostly Harmless Econometrics: An Empiricist’s Companion. Princeton: Princeton University Press.

Cameron, A. C., & Miller, D. L. 2015. A practitioner’s guide to cluster-robust inference. Journal of human resources, 50(2), 317-372.

Correia, S. 2016. reghdfe: Estimating linear models with multi-way fixed effects. In 2016 Stata Conference (No. 24). Stata Users Group.

Correia, S. 2019. REGHDFE: Stata module to perform linear or instrumental-variable regression absorbing any number of high-dimensional fixed effects.

Davidson, R., & MacKinnon, J. G. 2010. Wild bootstrap tests for IV regression. Journal of Business & Economic Statistics, 28(1), 128-144.

Roodman, D., Nielsen, M. Ø., MacKinnon, J. G., & Webb, M. D. 2019. Fast and wild: Bootstrap inference in Stata using boottest. The Stata Journal, 19(1), 4-60.

Sanderson, E. and F. Windmeijer, 2015. A Weak Instrument F-Test in Linear IV Models with Multiple Endogenous Variables. Journal of Econometrics
}

\keyword{ package }

\seealso{
    \code{\link{fastplm}}
}

