\name{summary.fastplm}
\alias{summary.fastplm}
\title{Print Estimated Results of Fixed Effect Model}

\description{Print estimated results of fixed effect model.}

\usage{\method{summary}{fastplm}(object, ...)}

\arguments{
  \item{object}{An \code{fastplm} object.}
  
  \item{...}{Other arguments.}
}

%\author{}
