\name{summary.fastplm}
\alias{summary.fastplm}
\title{Print Estimated Results of Fixed Effect Model}

\description{Print estimated results of fixed effect model.}

\usage{\method{summary}{fastplm}(object, detail = FALSE ,
  \dots)
}

\arguments{
  \item{object}{An \code{fastplm} object.}
  \item{detail}{A logical flag indicating whether to report diagnostic results for 2sls regressions.
	Default to FALSE.}
  \item{...}{Other arguments.}
}


